@@ -31,15 +31,15 @@
     "testEnvironment": "node"
   },
   "dependencies": {
-    "@typegoose/typegoose": "^12.1.0",
+    "@typegoose/typegoose": "^12.2.0",
     "bcryptjs": "^2.4.3",
     "connect-redis": "^7.1.1",
     "connection-string": "^4.4.0",
     "cors": "^2.8.5",
     "date-fns": "^3.3.1",
-    "dotenv": "^16.4.1",
-    "exifreader": "^4.20.0",
-    "express": "^4.18.2",
+    "dotenv": "^16.4.5",
+    "exifreader": "^4.21.0",
+    "express": "^4.18.3",
     "express-session": "^1.18.0",
     "fs-extra": "^11.2.0",
     "helmet": "^7.1.0",
@@ -48,12 +48,12 @@
     "p-limit": "~3.1.0",
     "passport": "^0.7.0",
     "passport-local": "^1.0.0",
-    "pino": "^8.17.2",
+    "pino": "^8.19.0",
     "pino-pretty": "^10.3.1",
     "pm2": "^5.3.1",
     "promisify-child-process": "^4.1.2",
     "readline-sync": "^1.4.10",
-    "redis": "^4.6.12",
+    "redis": "^4.6.13",
     "reflect-metadata": "^0.2.1",
     "sharp": "0.32.6",
     "superagent": "^8.1.2",
@@ -64,39 +64,34 @@
     "zod": "^3.22.4"
   },
   "devDependencies": {
-    "@commitlint/cli": "^18.6.0",
-    "@commitlint/config-conventional": "^18.6.0",
-    "@semantic-release/changelog": "^6.0.3",
-    "@semantic-release/git": "^10.0.1",
-    "@semantic-release/npm": "^11.0.2",
+    "@commitlint/cli": "^19.0.3",
+    "@commitlint/config-conventional": "^19.0.3",
     "@tsconfig/node20": "^20.1.2",
     "@types/bcryptjs": "^2.4.6",
     "@types/cors": "^2.8.17",
     "@types/express": "^4.17.21",
-    "@types/express-session": "^1.17.10",
+    "@types/express-session": "^1.18.0",
     "@types/fs-extra": "^11.0.4",
-    "@types/jest": "^29.5.11",
+    "@types/jest": "^29.5.12",
     "@types/passport": "^1.0.16",
     "@types/passport-local": "^1.0.38",
     "@types/readline-sync": "^1.4.8",
     "@types/supertest": "^6.0.2",
-    "@typescript-eslint/eslint-plugin": "^6.20",
-    "@typescript-eslint/parser": "^6.20",
     "conventional-changelog-conventionalcommits": "^7.0.2",
-    "eslint": "^8.56.0",
+    "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-import-resolver-typescript": "^3.6.1",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-no-autofix": "^1.2.3",
     "eslint-plugin-promise": "^6.1.1",
-    "husky": "^9.0.7",
+    "husky": "^9.0.11",
     "jest": "^29.7.0",
     "mongodb-memory-server": "^9.1.6",
-    "prettier": "3.2.4",
-    "semantic-release-export-data": "^1.0.1",
+    "prettier": "3.2.5",
     "supertest": "^6.3.4",
     "ts-jest": "^29.1.2",
     "ts-node-dev": "^2.0.0",
-    "typescript": "^5.3.3"
+    "typescript": "^5.3.3",
+    "typescript-eslint": "^7.2.0"
   }
 }
