@@ -1,7 +1,7 @@
 import { execFile } from "promisify-child-process";
 import fs from "fs-extra";
 import tempfile from "tempfile";
-import { logger } from ".";
+import { logger } from "./utils";
 
 export const updateImageDescription = async (
   file: string,
@@ -17,7 +17,6 @@ export const updateImageDescription = async (
   await fs.writeFile(argFile, description);
   const newFile = tempfile(".newFile");
   await fs.copy(file, newFile, {
-    overwrite: false,
     errorOnExist: true,
     preserveTimestamps: true,
   });
